requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.30.1
python-multipart==0.0.9
openai==1.40.2
pydantic==2.8.2
python-dotenv==1.0.0